Add files via upload
This header file provides a collection of functions and utilities to enhance the visual appeal and aesthetic of project. 
By incorporating this header file into project, you can easily change the color scheme and make your project more attractive to users.
Feel free to suggest additional functions or improvements to further enhance the visual appeal and aesthetic of projects.